Updates based on feedback from PR.
Motivation:

Address feedback.

Modifications:

New TestData file and modifications to example.

Result:

More concise example.
